+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Settings screen does now have a button to remove all links between Devices and Ports
   * Bugfix: Commiting a single Switch now leaves out the switchlinks-integrity check, which makes it possible to commit without all Switches running
   * New Setting-Element to unify settings handling
+    * Moved all settings and config variables to this Element
     * New corresponding Endpoint, that is now use by Frontend
     * New section in Settings-Screen to be able to modify all settings
   * Reworked System-Endpoint to not expose settings indirectly